requirements.txt
astroid==1.6.1
click==6.7
colorama==0.3.9
coverage==4.5.1
flasgger==0.8.1
Flask==0.12.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nose==1.3.7
PyJWT==1.5.3
pylint==1.8.2
PyYAML==3.12
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
